remove old cpp
cleanup includes
